:+1:  Smooth Motion
The code successfully generates some smooth motion, but not as expected with the Jacobian. The psuedo Jacobian also turns out to be more difficult than we thought (or I'm doing something wrong). The leg is also moving in a weird direction. It doesn't follow a path to a point. It just circles.